Cleanup in hdc_fortran

+++ examples/cpo_hdc.f90
@@ -12,7 +12,7 @@ program cpo_hdc
 contains
 
     subroutine test_cpos(equilibriumin, distsourceout)
-        integer(kind=8) :: i
+        integer(kind=c_size_t) :: i
         !UAL ! Always describe cpo as array
         !UAL ! In case of time slice, the size of the input cpo is 1
         !UAL type (type_equilibrium),pointer :: equilibriumin(:)
